PROJ-118 refactor: change file names, delete unnecessary code
