+++ features/avatarImage/index.ts
@@ -1,1 +1,2 @@
 export * from './ui'
+export * from './api'
