+++ features/avatarImage/ui/imageCropper/ImageCropper.tsx
@@ -1,3 +1,4 @@
+'use client'
 import React from 'react'
 import { useDropzone } from 'react-dropzone'
 import { ReactCrop } from 'react-image-crop'
