+++ features/avatarImage/ui/index.ts
@@ -1,1 +1,2 @@
 export * from './AvatarLoader'
+export * from './imageCropper'
