+++ features/createPost/ui/steps/Publication/PublicationContainer.tsx
@@ -8,27 +8,25 @@ import s from './PublicationContainer.module.scss'
 import { LocationPost } from './LocationPost/LocationPost'
 
 export const PublicationContainer = () => {
-  {
-    const { description, images } = useAppSelector(state => state.createPost.currentPost)
-    const filteredImages = useMemo(() => images.map(image => image.filteredImage), [images])
+  const { description, images } = useAppSelector(state => state.createPost.currentPost)
+  const filteredImages = useMemo(() => images.map(image => image.filteredImage), [images])
 
-    const dispatch = useAppDispatch()
+  const dispatch = useAppDispatch()
 
-    const handleDescriptionChange = (newDescription: string) => {
-      dispatch(updateDescription({ newDescription }))
-    }
-
-    return (
-      <div className={s.container}>
-        <Carousel images={filteredImages} />
-        <DescriptionPost
-          description={description}
-          sendNewPostDescription={handleDescriptionChange}
-          showSeparator
-        >
-          <LocationPost />
-        </DescriptionPost>
-      </div>
-    )
+  const handleDescriptionChange = (newDescription: string) => {
+    dispatch(updateDescription({ newDescription }))
   }
+
+  return (
+    <div className={s.container}>
+      <Carousel images={filteredImages} />
+      <DescriptionPost
+        description={description}
+        sendNewPostDescription={handleDescriptionChange}
+        showSeparator
+      >
+        <LocationPost />
+      </DescriptionPost>
+    </div>
+  )
 }
