+++ features/publicPage/ui/Posts/Post/PostSkeleton/index.ts
