+++ features/publicPage/ui/Posts/Post/index.ts
@@ -1,2 +1,2 @@
 export * from './Post'
-export * from './Skeleton'
+export * from './PostSkeleton'
