+++ features/publicPage/ui/Posts/PostsSkeleton/PostsSkeleton.module.scss
