+++ features/publicPage/ui/Posts/PostsSkeleton/index.ts
