+++ features/publicPage/ui/Posts/index.ts
@@ -1,3 +1,3 @@
 export * from './Post'
 export * from './Posts'
-export * from './Skeleton'
+export * from './PostsSkeleton'
