+++ features/publicPage/ui/RegisteredUsers/RegisteredUsersSkeleton/RegisteredUsersSkeleton.tsx
