+++ features/publicPage/ui/index.ts
@@ -1,1 +1,3 @@
 export * from './PublicPage'
+export * from './Posts'
+export * from './RegisteredUsers'
